Game logic worked on
Worked on getting the main game locic working.  Can click on squares
and have x and o's appear on the screen.  Will look for winner
and display the winner on the top of the screen.  Also has a reset
button that works.

Need to make it so after a winner is declared the users can't
make any more selections.